+++ staticsite/contents.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from typing import Optional, Dict, List, Any
+from typing import Dict, List, Any
 from .utils import front_matter, open_dir_fd
 from .page_filter import compile_page_match
 from . import site
@@ -15,14 +15,15 @@ class BaseDir:
     """
     Base class for content loaders
     """
-    def __init__(self, site: "site.Site", tree_root: str, relpath: str, dir_fd: int):
+    def __init__(self, site: "site.Site", tree_root: str, relpath: str, dir_fd: int, meta: Dict[str, Any]):
         self.site = site
         self.tree_root = tree_root
         self.relpath = relpath
         self.dir_fd = dir_fd
         self.subdirs: List[str] = []
         self.files: Dict[str, file.File] = {}
-        self.meta: Optional[Dict[str, Any]] = None
+        self.meta: Dict[str, Any] = meta
+        self.config: Dict[str, Any] = {}
 
         # Scan directory contents
         with os.scandir(self.dir_fd) as entries:
@@ -38,7 +39,7 @@ class BaseDir:
                     # Load .staticsite if found
                     with open(entry.name, "rt", opener=self._file_opener) as fd:
                         lines = [line.rstrip() for line in fd]
-                        fmt, self.meta = front_matter.parse(lines)
+                        fmt, self.config = front_matter.parse(lines)
                 elif entry.name.startswith("."):
                     # Skip hidden files
                     continue
@@ -51,32 +52,36 @@ class BaseDir:
                             abspath=os.path.join(self.tree_root, relpath),
                             stat=entry.stat())
 
+        # Read site metadata
+        if "site" in self.config:
+            self.meta.update(self.config["site"])
+
         # Postprocess directory metadata
         self.file_meta: Dict[str, Any] = {}
-        if self.meta is None:
-            self.meta = {}
-        else:
-            # Compute metadata for files
-            file_meta = self.meta.get("files")
-            if file_meta is not None:
-                rules = [(compile_page_match(k), v) for k, v in file_meta.items()]
-                for fname in self.files.keys():
-                    res: Dict[str, Any] = {}
-                    for pattern, meta in rules:
-                        if pattern.match(fname):
-                            res.update(meta)
-                    self.file_meta[fname] = res
 
-            # Compute metadata for directories
-            dir_meta = self.meta.get("dirs")
-            if dir_meta is not None:
-                rules = [(compile_page_match(k), v) for k, v in dir_meta.items()]
-                for dname in self.subdirs:
-                    res = {}
-                    for pattern, meta in rules:
-                        if pattern.match(dname):
-                            res.update(meta)
-                    self.file_meta[dname] = res
+        # Compute metadata for files
+        file_meta = self.config.get("files")
+        if file_meta is None:
+            file_meta = {}
+        rules = [(compile_page_match(k), v) for k, v in file_meta.items()]
+        for fname in self.files.keys():
+            res: Dict[str, Any] = dict(self.meta)
+            for pattern, meta in rules:
+                if pattern.match(fname):
+                    res.update(meta)
+            self.file_meta[fname] = res
+
+        # Compute metadata for directories
+        dir_meta = self.config.get("dirs")
+        if dir_meta is None:
+            dir_meta = {}
+        rules = [(compile_page_match(k), v) for k, v in dir_meta.items()]
+        for dname in self.subdirs:
+            res: Dict[str, Any] = dict(self.meta)
+            for pattern, meta in rules:
+                if pattern.match(dname):
+                    res.update(meta)
+            self.file_meta[dname] = res
 
     def meta_file(self, fname: str):
         res = self.file_meta.get(fname)
@@ -101,19 +106,24 @@ class ContentDir(BaseDir):
 
         log.debug("Loading pages from %s:%s", self.tree_root, self.relpath)
 
+        # Store directory metadata
+        self.site.dirs[self.relpath] = self.meta
+
         # Load subdirectories
         for fname in self.subdirs:
             # Check whether to load subdirectories as asset trees
             meta = self.file_meta.get(fname)
             if meta and meta.get("asset"):
                 with open_dir_fd(fname, dir_fd=self.dir_fd) as subdir_fd:
-                    subdir = AssetDir(self.site, self.tree_root, os.path.join(self.relpath, fname), subdir_fd)
+                    subdir = AssetDir(
+                                self.site, self.tree_root, os.path.join(self.relpath, fname), subdir_fd, meta=meta)
                     subdir.load()
             else:
                 # TODO: prevent loops with a set of seen directory devs/inodes
                 # Recurse
                 with open_dir_fd(fname, dir_fd=self.dir_fd) as subdir_fd:
-                    subdir = ContentDir(self.site, self.tree_root, os.path.join(self.relpath, fname), subdir_fd)
+                    subdir = ContentDir(
+                                self.site, self.tree_root, os.path.join(self.relpath, fname), subdir_fd, meta=meta)
                     subdir.load()
 
         # Handle files marked as assets in their metadata
@@ -163,24 +173,29 @@ class AssetDir(BaseDir):
 
         log.debug("Loading pages from %s:%s", self.tree_root, self.relpath)
 
+        # Store directory metadata
+        self.site.dirs[self.relpath] = self.meta
+
         # Load subdirectories
         for fname in self.subdirs:
             # TODO: prevent loops with a set of seen directory devs/inodes
             # Recurse
+            meta = self.file_meta.get(fname)
             with open_dir_fd(fname, dir_fd=self.dir_fd) as subdir_fd:
                 subdir = AssetDir(
                             self.site,
                             self.tree_root,
                             os.path.join(self.relpath, fname),
                             subdir_fd,
-                            dest_subdir=self.dest_subdir)
+                            dest_subdir=self.dest_subdir, meta=meta)
                 subdir.load()
 
         # Use everything else as an asset
         for fname, f in self.files.items():
             if stat.S_ISREG(f.stat.st_mode):
                 log.debug("Loading static file %s", f.relpath)
-                p = Asset(self.site, f, meta=self.file_meta.get(fname), dest_subdir=self.dest_subdir)
+                meta = self.file_meta.get(fname)
+                p = Asset(self.site, f, dest_subdir=self.dest_subdir, meta=meta)
                 if not p.is_valid():
                     continue
                 self.site.add_page(p)